ai/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.5
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.5
bitsandbytes==0.43.2
black==24.10.0
bleach==6.2.0
blinker==1.9.0
build==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.1.0
cleo==2.1.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
crashtest==0.4.1
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.70.0
datasets==4.2.0
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.4.0
distlib==0.4.0
distro==1.9.0
Django==4.2.21
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
docker==7.1.0
dotenv==0.9.9
drf-yasg==1.21.7
dulwich==0.24.5
emoji==2.15.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.120.0
fastjsonschema==2.21.2
filelock==3.13.1
findpython==0.7.0
flake8==7.1.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.73.1
h11==0.14.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.35.3
idna==3.4
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
installer==0.7.0
ipykernel==6.30.1
ipython==8.37.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.5.1
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.11.3
keyring==25.6.0
kiwisolver==1.4.9
langchain==1.0.4
langchain-core==1.0.3
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lark==1.2.2
libclang==18.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.4
ml_dtypes==0.5.3
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
openai==1.70.0
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pbs-installer==2025.10.14
peft==0.10.0
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
prometheus_client==0.22.1
prompt-toolkit==3.0.43
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.11.2
pydantic_core==2.33.1
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2023.3.post1
pywin32==310
pywin32-ctypes==0.2.3
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==26.3.0
qtconsole==5.7.0
QtPy==2.4.3
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.15.3
Send2Trash==1.8.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.15.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
soynlp==0.0.493
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.20.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.1
transformers==4.57.1
trl==0.24.0
trove-classifiers==2025.9.11.17
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.38.0
validators==0.35.0
virtualenv==20.35.3
waitress==3.0.2
wcwidth==0.2.13
weaviate-client==4.17.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.21.0
zstandard==0.25.0
